requirements.txt
fastapi
uvicorn
sounddevice
numpy
pyaudio
faster-whisper
edge-tts
google-genai
beautifulsoup4
playwright
httpx
mcp
python-dotenv
pydantic
